fix(incomplete): correct styles contact form ads
